requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==6.0.0
    # via streamlit
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.6
    # via streamlit
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via
    #   reportlab
    #   requests
click==8.3.1
    # via streamlit
colorama==0.4.6
    # via click
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
idna==3.11
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.15.0
    # via altair
numpy==2.4.1
    # via
    #   opencv-python
    #   pandas
    #   pydeck
    #   streamlit
opencv-python==4.13.0.90
    # via image2pdf-without-ocr (pyproject.toml)
packaging==26.0
    # via
    #   altair
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==12.1.0
    # via
    #   reportlab
    #   streamlit
protobuf==6.33.5
    # via streamlit
pyarrow==23.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
reportlab==4.4.9
    # via image2pdf-without-ocr (pyproject.toml)
requests==2.32.5
    # via streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.53.1
    # via image2pdf-without-ocr (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   referencing
    #   streamlit
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
watchdog==6.0.0
    # via streamlit
